+++ core/src/main/java/ru/tinkoff/kotea/core/UncaughtExceptionHandler.kt
@@ -1,6 +1,0 @@
-package ru.tinkoff.kotea.core
-
-fun interface UncaughtExceptionHandler {
-
-    fun handle(throwable: Throwable)
-}