+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## 1.0.1
 
+* remove unhandled exception handler in store
+
 ## 1.0.0
 
 * First version